# Speedup hashmap by speeding up node creation