During profiling, I noticed a lot of time was being spent in memcopy. I tracked it down to less than ideal Node initialization. I ended up utilizing a combination of `UnsafeCell` and `MaybeUninit` to avoid utilizing unstable APIs.


These results are for small pairs `u64, u64` but the delta increases with larger items.

Before

```
test hashmap_insert_mut_10     ... bench:         407 ns/iter (+/- 5)
test hashmap_insert_mut_100    ... bench:       7,380 ns/iter (+/- 122)
test hashmap_insert_mut_1000   ... bench:      81,006 ns/iter (+/- 1,841)
test hashmap_insert_mut_10000  ... bench:     879,453 ns/iter (+/- 29,964)
```

After

```
test hashmap_insert_mut_10     ... bench:         353 ns/iter (+/- 7)
test hashmap_insert_mut_100    ... bench:       5,844 ns/iter (+/- 87)
test hashmap_insert_mut_1000   ... bench:      63,217 ns/iter (+/- 7,667)
test hashmap_insert_mut_10000  ... bench:     738,824 ns/iter (+/- 28,214)
```