@jneem 2024-04-01
Thanks for the PR! This looks fine to me, but I'm going to finish running tests in miri just in case.

It does make me wonder whether it's worth switching to `triomphe` unconditionally, though. I was able to get similar performance improvements with much simpler unsafe code when using `triomphe::UniqueArc`.