@arthurprs 2024-04-02
Yeah, things like [UniqueArc::new_uninit](https://docs.rs/triomphe/latest/triomphe/struct.UniqueArc.html#method.new_uninit) could greatly simplify things. At that point we are probably also removing the entire Pool thingy, which is a much larger PR.